requirements.txt
Django==4.0.5
django-extensions==3.1.5
django-filter==21.1
dj-database-url==0.5.0
python-decouple==3.6
